initial checkin, basic functionality
Everything in one file, except for the heart image that is not
included. No real error and browser support checking, developed to
run in firefox. Comments are german (might change later on).

The website draws a static background to the canvas element and
then starts adding dynamic content, like a sun and beams, clouds
and of course the thing all this is named after: a tree.

Up to to:
- random colored strips at random leaf locations
  - make them bend in the wind
  - animate them maybe
- make the tree look more like a birch (black outline, white main
  color and some random looking horizontal black strips)
- maybe add some grass? Birds? Butterflies?